# SwaggerOffLine.yaml
openapi: 3.1.0
info:
  title: Auto Escola
  description: API REST para gerenciamento de instruções de uma autoescola.
  contact:
    name: Equipe de desenvolvimento
  license:
    name: Apache 2.0
    url: http://springdoc.org
servers:
  - url: http://localhost:8085
    description: Generated server url
paths:
...
  /login:
    post:
      tags:
        - login-controller
      operationId: efetuarLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosLogin"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string
...
components:
  schemas:
...
    DadosLogin:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
      required:
        - login
        - senha
...
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT