# src/main/java/br/com/autoescola/api/adapter/in/controller/request/usuario/DadosLogin.java
package br.com.autoescola.api.adapter.in.controller.request.usuario;

import jakarta.validation.constraints.NotBlank;

public record DadosLogin(
        @NotBlank
        String login,

        @NotBlank
        String senha) {
}